requirements.txt
# requirements.txt
numpy==1.26.4
opencv-python>=4.5.0
albumentations>=1.0.0
pycocotools>=2.0.0
pyyaml>=5.4.0

--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
pandas
torchmetrics
ultralytics